requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-community>=0.0.20

# LLM Providers
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langchain-groq>=0.0.1
langchain-anthropic>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Vector Store
chromadb>=0.4.0

# Environment and Configuration
python-dotenv>=1.0.0

# Database
sqlite3  # Built-in to Python

# Graph Visualization
graphviz>=0.20.0
pygraphviz>=1.11  # Optional

# Optional: External Search APIs
# tavily-python>=0.2.0
# google-search-results>=2.4.0
